Fixed int / longs to decimals

+++ Hexio.DineroClient/Models/Products/CreateProductModel.cs
@@ -4,8 +4,8 @@ namespace Hexio.DineroClient.Models.Products
     {
         public string ProductNumber { get; set; }
         public string Name { get; set; }
-        public long? BaseAmountValue { get; set; }
-        public long? Quantity { get; set; }
+        public decimal? BaseAmountValue { get; set; }
+        public decimal? Quantity { get; set; }
         public long AccountNumber { get; set; }
         public string Unit { get; set; }
         public string ExternalReference { get; set; }
